+++ src/main/java/com/hhplu/hhplusconcert/app/application/facade/WaitingQueueFacade.java
@@ -1,6 +1,6 @@
 package com.hhplu.hhplusconcert.app.application.facade;
 
-import com.hhplu.hhplusconcert.app.application.service.UserService;
+import com.hhplu.hhplusconcert.app.application.user.service.UserService;
 import com.hhplu.hhplusconcert.app.application.waitingqueue.service.WaitingQueueService;
 import com.hhplu.hhplusconcert.app.application.waitingqueue.command.CreateWaitingQueueCommand;
 import com.hhplu.hhplusconcert.app.application.waitingqueue.command.GetWaitingQueueCommand;
