+++ src/main/java/com/hhplu/hhplusconcert/app/interfaces/api/concert/dto/SeatItem.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Getter
 @Builder
 @AllArgsConstructor
-public class SeatValue {
+public class SeatItem {
     private Long seatId;
     private Long seatNumber;
     private String seatStatus;
