+++ src/main/java/com/hhplu/hhplusconcert/app/interfaces/api/concert/res/ScheduleResponse.java
@@ -1,7 +1,6 @@
 package com.hhplu.hhplusconcert.app.interfaces.api.concert.res;
 
 import com.hhplu.hhplusconcert.app.application.concert.command.ConcertResponseCommand;
-import com.hhplu.hhplusconcert.app.interfaces.api.concert.dto.SeatValue;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
