+++ src/test/java/com/hhplu/hhplusconcert/app/application/facade/ReservationFacadeTest.java
@@ -2,10 +2,10 @@ package com.hhplu.hhplusconcert.app.application.facade;
 
 import com.hhplu.hhplusconcert.app.application.reservation.command.ReserveSeatsCommand;
 import com.hhplu.hhplusconcert.app.application.reservation.command.ReserveSeatsResponseCommand;
-import com.hhplu.hhplusconcert.app.application.service.ConcertService;
-import com.hhplu.hhplusconcert.app.application.service.PaymentService;
+import com.hhplu.hhplusconcert.app.application.concert.service.ConcertService;
+import com.hhplu.hhplusconcert.app.application.payment.service.PaymentService;
 import com.hhplu.hhplusconcert.app.application.reservation.service.ReservationService;
-import com.hhplu.hhplusconcert.app.application.service.SeatService;
+import com.hhplu.hhplusconcert.app.application.concert.service.SeatService;
 import com.hhplu.hhplusconcert.app.domain.concert.entity.Seat;
 import com.hhplu.hhplusconcert.app.domain.reservation.entity.Reservation;
 import jakarta.persistence.OptimisticLockException;
