+++ src/test/java/com/hhplu/hhplusconcert/app/application/token/SeatServiceTest.java
@@ -1,5 +1,6 @@
-package com.hhplu.hhplusconcert.app.application.service;
+package com.hhplu.hhplusconcert.app.application.token;
 
+import com.hhplu.hhplusconcert.app.application.concert.service.SeatService;
 import com.hhplu.hhplusconcert.app.domain.concert.SeatStatus;
 import com.hhplu.hhplusconcert.app.domain.concert.entity.Seat;
 import com.hhplu.hhplusconcert.app.domain.concert.repository.SeatRepository;
